Fix: time remaining now updated every interval of fixed period (1 second)
As checked, measured execution time between beginning and ending inside OnTimer() is pretty
much not waste any time at all (~0). Thus it is very fast, so no issue there.

But still the issue keeps happening for lagging of time remaining doesn't keep up with the
current calculated time. Forcing chart redraw is usually called as mentioned in the document
after chaning object's properties, but still then it didn't work perfectly for us.
Thus explicitly call to force it is a way to fix an issue here.